+++ CHANGELOG.md
@@ -5,6 +5,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.6.13] - 2022-04-04
+
 - `AttributesParameter` class:
   - `ATTRIBUTES_PARAMETERS` has been added (as an array that contains `AttributesParameters` classes)
 - `AttributesParameters` class has been added:
@@ -252,7 +254,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - CatalogGraphQLClient class with its "find" method
 
-[Unreleased]: https://github.com/ArrowSphere/catalog-graphql-client/compare/0.6.12...HEAD
+[Unreleased]: https://github.com/ArrowSphere/catalog-graphql-client/compare/0.6.13...HEAD
+[0.6.13]: https://github.com/ArrowSphere/catalog-graphql-client/compare/0.6.12...0.6.13
 [0.6.12]: https://github.com/ArrowSphere/catalog-graphql-client/compare/0.6.11...0.6.12
 [0.6.11]: https://github.com/ArrowSphere/catalog-graphql-client/compare/0.6.10...0.6.11
 [0.6.10]: https://github.com/ArrowSphere/catalog-graphql-client/compare/0.6.9...0.6.10